+++ R/progressr_uniroot.R
@@ -5,9 +5,10 @@ uniroot_with_progress <- function(
   abs_tol = 1e-6,
   maxiter = 100
 ) {
-  last_val <- NA
-  first_interval <- NA
-  last_reported_progress <- 0
+  state <- new.env(parent = emptyenv())
+  state$last_val <- NA
+  state$first_interval <- NA
+  state$last_reported_progress <- 0
 
   left_val_dist <- fun(bounds[1]) - target
   right_val_dist <- fun(bounds[2]) - target
@@ -22,17 +23,17 @@ uniroot_with_progress <- function(
     logger::log_debug("Evaluating at ", target_val)
     est <- fun(target_val)
     error <- est - target
-    if (!is.na(last_val)) {
-      if (est - last_val == 0) {
+    if (!is.na(state$last_val)) {
+      if (est - state$last_val == 0) {
         progress <- 1
       } else {
         conv_crit <- .Machine$double.eps * abs(target_val) + abs_tol / 2
-        interval <- abs(last_val - target_val)
-        if (is.na(first_interval)) {
-          first_interval <<- interval
+        interval <- abs(state$last_val - target_val)
+        if (is.na(state$first_interval)) {
+          state$first_interval <- interval
         }
-        progress <- (log(interval) - log(first_interval)) /
-          (log(conv_crit) - log(first_interval))
+        progress <- (log(interval) - log(state$first_interval)) /
+          (log(conv_crit) - log(state$first_interval))
       }
       logger::log_debug(
         "Current estimate: ", est,
@@ -40,7 +41,7 @@ uniroot_with_progress <- function(
         ", progress: ", progress
       )
       progress_percent <- min(ceiling(100 * progress), 100)
-      add_progress <- progress_percent - last_reported_progress
+      add_progress <- progress_percent - state$last_reported_progress
       if (add_progress > 0) {
         message <- paste(
           "Est.: ", round(est, 4), "(", round(target_val, 4),
@@ -49,7 +50,7 @@ uniroot_with_progress <- function(
         p(message, amount = add_progress)
       }
     }
-    last_val <<- target_val
+    state$last_val <- target_val
     error
   }
 
@@ -60,9 +61,10 @@ uniroot_with_progress <- function(
     maxiter = maxiter
   )
 
-  if (is.na(last_reported_progress) || last_reported_progress < 100) {
+  last_p <- state$last_reported_progress
+  if (is.na(last_p) || last_p < 100) {
     p(step = 100)
   }
 
-  return(results)
+  results
 }