+++ NEWS.md
@@ -1,6 +1,38 @@
 # CensRCalc (development version)
 
-* Initial package functionality
+Release candidate of 0.1.0.
+
+## Major changes
+
+* Added parameter `tau`, which allows evaluating the censoring proportion at a
+  specific time point $\tau$.
+* Added `plot()` and `print()` methods for the expected censoring proportion
+  function.
+* Added tests.
+* Better handling and guarantee for `abs_tol` for numerical integration.
+
+## Minor improvements and bug fixes
+
+* Added evaluation of the passed event survival function, using the marginal.
+* Removed renv usage.
+* Development workflow improvements and fixes.
+* Documentation improvements and fixes.
+
+# CensRCalc 0.1.0
+
+* Added parameter `tau`, which allows evaluating the censoring proportion at a
+  specific time point $\tau$.
+* Added `plot()` and `print()` methods for the expected censoring proportion
+  function.
+* Added tests.
+* Better handling and guarantee for `abs_tol` for numerical integration.
+
+## Minor improvements and bug fixes
+
+* Added evaluation of the passed event survival function, using the marginal.
+* Removed renv usage.
+* Development workflow improvements and fixes.
+* Documentation improvements and fixes.
 
 # CensRCalc 0.0.1